refactor(BE): server.js 리팩토링
- 기존의 server.js에서 코드들을 분리해 controller.js와 model.js 생성
- controller.js: 비즈니스 로직 코드
- model.js: MongoDB 연결 설정 및 스키마 정의
